Refactoring of the code in the service post file and changes in the error warnings

+++ api/src/post/post.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Post } from './post.entity';
@@ -7,6 +7,7 @@ import { CreatePostDto } from './dto/createPost.dto';
 import { UserService } from '../user/user.service';
 import { UptadePostDto } from './dto/uptadePost.dto';
 import { title } from 'process';
+import { error } from 'console';
 
 @Injectable()
 export class PostService {
@@ -19,11 +20,13 @@ export class PostService {
 
   async getAllPosts(): Promise<Post[]> {
     const allPosts = await this.postRepository.find();
+    if (!allPosts) throw new NotFoundException('No se ha podido traer todos los post');
     return allPosts;
   }
 
   async getPostById(id: number): Promise<Post> {
     const findedPost = await this.postRepository.findOneByOrFail({ id });
+    if (!findedPost) throw new NotFoundException('No se ha podido traer todos los post');
     return findedPost;
   }
 
@@ -33,21 +36,28 @@ export class PostService {
     imageDataArray: Buffer[],
   ): Promise<Post> {
     const user = await this.userService.getUser({ id: userId });
+    if (!user) throw new NotFoundException('No se ha encontrado el usuario');
     const date: Date = new Date();
+    if (!date) throw new BadRequestException('No se ha podido crear la fecha');
     if (
       postDto.price !== undefined &&
       (postDto.price < 1 || postDto.price > 1000000)
     ) {
-      throw new Error(`El precio debe estar entre ${1} y ${1000000}.`);
+      throw new BadRequestException(`El precio debe estar entre ${1} y ${1000000}.`);
     }
     const newPost = this.postRepository.create(postDto);
     newPost.creationDate = date;
     newPost.user = user;
-    await this.postRepository.save(newPost);
+    if (!newPost) throw new BadRequestException('No se ha podido crear el post');
+    const savePost = await this.postRepository.save(newPost);
+    if (!savePost) throw new BadRequestException('No se ha podido guardar el post creado');
     const imagePosts = imageDataArray.map((imageData) =>
       this.imagePostRepository.create({ imageData, post: newPost }),
     );
-    await this.imagePostRepository.save(imagePosts);
+    if (!imagePosts) throw new BadRequestException('No se ha podido crear la imagen');
+    const saveImgPost = await this.imagePostRepository.save(imagePosts);
+    if (!saveImgPost)
+      throw new BadRequestException('No se ha podido guardar la imagen creada');
     return newPost;
   }
 
@@ -56,12 +66,18 @@ export class PostService {
     imageDataArray: Buffer[],
   ): Promise<Post> {
     const post = await this.getPostById(postId);
+    if (!post) throw new NotFoundException('No se ha encontrado el post');
     const imagePosts = imageDataArray.map((imageData) =>
       this.imagePostRepository.create({ imageData, post }),
     );
-    await this.imagePostRepository.save(imagePosts);
+    if (!imagePosts) throw new BadRequestException('No se ha podido crear las imagenes');
+    const saveImgPost = await this.imagePostRepository.save(imagePosts);
+    if (!saveImgPost)
+      throw new BadRequestException('No se ha podido guardar las imagenes creadas');
     post.images = [...post.images, ...imagePosts];
-    return await this.postRepository.save(post);
+    const saveAddImgPost = await this.postRepository.save(post);
+    if (!saveAddImgPost) throw new BadRequestException ('No se ha podido agregar imagenes al post')
+    return saveAddImgPost
   }
 
   async removeImageFromPost(
@@ -69,12 +85,17 @@ export class PostService {
     imagePostId: number,
   ): Promise<Post> {
     const post = await this.getPostById(postId);
+    if (!post) throw new NotFoundException('No se ha encontrado el post');
     const imagePost = await this.imagePostRepository.findOneOrFail({
       where: { id: imagePostId },
     });
     post.images = post.images.filter((image) => image.id !== imagePost.id);
-    await this.imagePostRepository.remove(imagePost);
-    return await this.postRepository.save(post);
+    if (!imagePost) throw new BadRequestException('No se han podido borrar las imagenes');
+    const ImgRemove = await this.imagePostRepository.remove(imagePost);
+    if (!ImgRemove) throw new BadRequestException('No se ha podido borrar las imagenes del post');
+    const ImgPostRemove = await this.postRepository.save(post);
+    if (!ImgPostRemove) throw new BadRequestException('No se ha podido borrar las imagenes guardadas del post');
+    return ImgPostRemove
   }
 
   async uptadePost(
@@ -82,23 +103,25 @@ export class PostService {
     uptadePostData: UptadePostDto,
   ): Promise<Post> {
     const postFound = await this.postRepository.findOneBy({ id: postId });
-    if (!postFound) throw new Error('La publicación no existe');
-
+    if (!postFound) throw new NotFoundException('La publicación no existe');
     const uptadePost = { ...postFound, ...this.uptadePost };
+    if(!uptadePost) throw new BadRequestException('No se ha podido actualizar la publicación')
     const savePost = await this.postRepository.save(uptadePost);
+    if(!savePost) throw new BadRequestException('No se ha podido guardar la actualizacion de la publicación')
     return savePost;
   }
 
   async deletePost(postId: number): Promise<undefined> {
     const postDelFound = await this.postRepository.findOneBy({ id: postId });
-    if (!postDelFound) throw new Error('La publicacion no existe');
-
-    await this.postRepository.delete(postId);
+    if (!postDelFound) throw new NotFoundException('La publicacion no existe');
+    const delPost = await this.postRepository.delete(postId);
+    if (!delPost) throw new NotFoundException('No se ha podido borrar la publicación');
     return undefined;
   }
 
   async getPostByName(title: string): Promise<Post> {
     const PostName = await this.postRepository.findOneBy({ title });
+    if (!PostName) throw new NotFoundException('No se ha podido encontrar la publicación')
     return PostName;
   }
 }
